requirements.txt
torch
numpy
Pillow
torchvision
scikit-learn 
matplotlib
seaborn